+++ s3tbx-sentinel3-reader/src/main/java/org/esa/s3tbx/dataio/s3/slstr/SlstrLevel1B500mProductFactory.java
@@ -43,8 +43,8 @@ public class SlstrLevel1B500mProductFactory extends SlstrLevel1ProductFactory {
         if(sourceBandNameLength > 1) {
             gridIndex = sourceBandName.substring(sourceBandNameLength - 2);
         }
-        final Integer sourceStartOffset = getStartOffset(gridIndex);
-        final Integer sourceTrackOffset = getTrackOffset(gridIndex);
+        final Double sourceStartOffset = getStartOffset(gridIndex);
+        final Double sourceTrackOffset = getTrackOffset(gridIndex);
         if (sourceStartOffset != null && sourceTrackOffset != null) {
             final short[] sourceResolutions = getResolutions(gridIndex);
             if (gridIndex.startsWith("t")) {
